// src/structures/types.d.ts
export declare type CommandArgument = {
    /**
     * The name of the argument.
     */
    trigger: string;
    /**
     * The description of the argument. Displayed in the help command and slash command preview.
     */
    description?: string;
    /**
     * How should the value be resolved or what should be prompted for slash commands?
     */
    type: 'Subcommand' | 'SubcommandGroup' | 'String' | 'Number' | 'Boolean' | 'Subcommand' | 'RobloxUser' | 'RobloxRole' | 'DiscordUser' | 'DiscordRole' | 'DiscordChannel' | 'DiscordMentionable';
    /**
     * Should the bot be sent requests to autocomplete everything they type?
     * @default false;
     */
    autocomplete?: boolean;
    /**
     * Is this argument required?
     */
    required?: boolean;
    /**
     * Is this argument inputted through a a flag when using the legacy commands system?
     * @default false
     */
    isLegacyFlag?: boolean;
    /**
     * The choices that the user can pick from.
     */
    choices?: ApplicationCommandOptionChoiceData<string>[];
    /**
     * If this is a subcommand (group), command arguments.
     */
    args?: CommandArgument[];
    /**
     * If the type of this argument is set to DiscordChannel, what channel types are allowed?
     */
    channelTypes?: ApplicationCommandOptionAllowedChannelTypes[];
}